+++ src/Components/CreatePoll.js
@@ -1,0 +1,13 @@
+import PollCreationPage from './PollCreationPage';
+import PollConfirmationPage from './PollConfirmationPage';
+
+function CreatePoll (){
+    return(
+        <>
+            <PollCreationPage />
+            <PollConfirmationPage />
+        </>
+    )
+}
+
+export default CreatePoll;
