--- src/Components/Footer.js
@@ -1,9 +1,0 @@
-function Footer() {
-  return (
-    <>
-    <footer>Hi I'm a foot.</footer>
-    </>
-  )
-}
-
-export default Footer;
