--- src/Components/NavBar.js
@@ -1,15 +1,0 @@
-function NavBar() {
-  return (
-    <>
-    <div className="navBar">
-      <ul>
-        <li>Go here</li>
-        <li>Go there</li>
-        <li>Go forth</li>
-      </ul>
-    </div>
-    </>
-  )
-}
-
-export default NavBar;
