--- src/Pages/CreatePoll.js
@@ -1,13 +1,0 @@
-import PollCreationPage from '../Components/PollCreation';
-import PollConfirmationPage from '../Components/PollConfirmationPage';
-
-function CreatePoll() {
-  return (
-    <>
-      <PollCreationPage />
-      <PollConfirmationPage />
-    </>
-  )
-}
-
-export default CreatePoll;
