--- src/Pages/ErrorPage.js
@@ -1,7 +1,0 @@
-function ErrorPage() {
-  return (
-    <h2>Oh no! Error!</h2>
-  )
-}
-
-export default ErrorPage;
