--- src/Pages/FindPoll.js
@@ -1,9 +1,0 @@
-function FindPoll() {
-  return (
-    <>
-    <h2>Find a Poll</h2>
-    </>
-  )
-};
-
-export default FindPoll;
