--- src/Pages/Home.js
@@ -1,8 +1,0 @@
-function Home() {
-  return (
-    <>
-    </>
-  )
-};
-
-export default Home;
