--- src/Pages/Results.js
@@ -1,8 +1,0 @@
-function Results() {
-  return (
-    <>
-    </>
-  )
-};
-
-export default Results;
